uv as uint8_t and limits in config

+++ arduino/arduSiser3/siser.h
@@ -4,7 +4,7 @@ struct myDht_t {
   float t;
   float h;
 };
-
+  
 struct myAcc_t {
   float x;
   float y;
@@ -14,7 +14,7 @@ struct myAcc_t {
 struct siser_t {
   uint8_t id;
   uint8_t rain;  // 0-100%
-  float uv;
+  uint8_t uv;
   myDht_t dht;
   float bat;     // battery: 0-5V
   myAcc_t acc;   // accelerometer
@@ -23,7 +23,7 @@ struct siser_t {
 siser_t siser = { // siser Example
   ID,        // id: 0 - 255
   15,        // rain
-  0.0,       // UV
+  0,         // UV
   {23.5,51}, // dht:t,h
   50.0,      // Battery voltage
   {0.1,0.2,-9.07}
